+++ R/modify.R
@@ -1,11 +1,17 @@
 
-#' Open the Lockfile for Editing
+#' Modify a Lockfile
 #'
-#' Open a project's lockfile (if any) for editing. After edit, if the lockfile
-#' edited is associated with the active project, any state-related changes
-#' (e.g. to \R repositories) will be updated in the current session.
+#' Modify a project's lockfile, either interactively or non-interactively.
+#'
+#' After edit, if the lockfile edited is associated with the active project, any
+#' state-related changes (e.g. to \R repositories) will be updated in the
+#' current session.
 #'
 #' @inherit renv-params
+#'
+#' @param changes A list of changes to be merged into the lockfile.
+#'   When `NULL` (the default), the lockfile is instead opened for
+#'   interactive editing.
 #'
 #' @export
 #'
@@ -17,31 +23,98 @@
 #'   renv::modify()
 #'
 #' }
-modify <- function(project = NULL) {
+modify <- function(project = NULL, changes = NULL) {
   renv_scope_error_handler()
   project <- renv_project_resolve(project)
   renv_scope_lock(project = project)
-  renv_modify_impl(project)
+  renv_modify_impl(project, changes)
   invisible(project)
 }
 
-renv_modify_impl <- function(project) {
+renv_modify_impl <- function(project, changes) {
 
-  path <- renv_lockfile_path(project)
-  if (!file.exists(path))
-    stopf("lockfile '%s' does not exist", aliased_path(path))
+  lockfile <- if (is.null(changes))
+    renv_modify_interactive(project)
+  else
+    renv_modify_noninteractive(project, changes)
 
-  renv_file_edit(path)
+  if (identical(renv_project(), project))
+    renv_modify_fini(lockfile)
 
-  if (!renv_path_same(project, renv_project()))
-    return(NULL)
+}
 
-  lockfile <- catch(renv_lockfile_load(project))
+renv_modify_interactive <- function(project) {
+
+  # check for interactive session
+  if (!interactive())
+    stop("can't modify lockfile in non-interactive session")
+
+  # resolve path to lockfile
+  lockpath <- renv_lockfile_path(project)
+  if (!file.exists(lockpath))
+    stopf("lockfile '%s' does not exist", aliased_path(lockpath))
+
+  # copy the lockfile to a temporary file
+  dir <- tempfile("renv-lockfile-")
+  ensure_directory(dir)
+  on.exit(unlink(dir, recursive = TRUE), add = TRUE)
+
+  templock <- file.path(dir, "renv.lock")
+  file.copy(lockpath, templock)
+
+  # edit the temporary lockfile
+  renv_file_edit(templock)
+
+  # check that the new lockfile can be read
+  lockfile <- catch(renv_lockfile_read(file = templock))
   if (inherits(lockfile, "error")) {
-    warning(lockfile)
-    return(NULL)
+
+    renv_pretty_print(
+      conditionMessage(lockfile),
+      preamble  = "renv was uanble to parse the modified lockfile:",
+      postamble = "Your changes will be discarded.",
+      wrap = FALSE
+    )
+
+    stop("error modifying lockfile")
+
   }
 
+  lockfile
+
+}
+
+renv_modify_noninteractive <- function(project, changes) {
+
+  # resolve path to lockfile
+  lockpath <- renv_lockfile_path(project)
+  if (!file.exists(lockpath))
+    stopf("lockfile '%s' does not exist", aliased_path(lockpath))
+
+  # read it
+  lockfile <- renv_lockfile_read(file = lockpath)
+
+  # merge changes
+  merged <- modifyList(lockfile, changes)
+
+  # write updated lockfile to a temporary file
+  templock <- tempfile("renv-lock-")
+  renv_lockfile_write(merged, file = templock)
+
+  # try reading it once more
+  renv_lockfile_read(file = templock)
+
+  # overwrite the original lockfile
+  file.rename(templock, lockpath)
+
+  # finish up
+  merged
+
+}
+
+renv_modify_fini <- function(lockfile) {
+
+  # synchronize relevant changes into the session
   repos <- lockfile$R$Repositories
   options(repos = convert(repos, "character"))